Move Level to rustc_session
