+++ src/librustc_session/lib.rs
@@ -1,2 +1,3 @@
 pub mod cgu_reuse_tracker;
 pub mod utils;
+pub mod lint;
